+++ lib/extraFunc.lua
@@ -70,6 +70,13 @@ function ReverseList(ls) --reverse a table of simple values
     return ls
 end
 
+function math.sign(x)
+    if x > 0 then return 1
+    elseif x < 0 then return -1
+    else return 0
+    end
+end
+
 local cachedFontSize = nil
 local cachedFontObject = nil